+++ cucumber-test/src/main/java/TestDeck.java
@@ -4,8 +4,8 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
-import cucumbertest.Robot;
 import decks.Deck;
+import model.tile.Robot;
 
 public class TestDeck {
