+++ cucumber-test/src/main/java/TestReplayCard.java
@@ -2,8 +2,8 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
-import cucumbertest.Robot;
 import decks.Deck;
+import model.tile.Robot;
 
 public class TestReplayCard {
