+++ cucumber-test/src/main/java/TestRotationRight.java
@@ -2,8 +2,8 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
-import cucumbertest.Robot;
 import decks.Deck;
+import model.tile.Robot;
 import subcards.RotateLeft;
 import subcards.RotateRight;
