--- cucumber-test/src/main/java/cucumbertest/BoardFactory.java
@@ -1,5 +1,0 @@
-package cucumbertest;
-
-public class BoardFactory {
-
-}
