+++ cucumber-test/src/main/java/subcards/MoveTwo.java
@@ -1,6 +1,6 @@
 package subcards;
 
-import cucumbertest.Robot;
+import model.tile.Robot;
 
 public class MoveTwo extends Card {
