+++ cucumber-test/src/test/java/StepsDefinition.java
@@ -1,8 +1,8 @@
 
-import cucumbertest.Robot;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
+import model.tile.Robot;
 
 public class StepsDefinition {
 	Robot robot = new Robot(0,0);
