+++ cucumber-test/src/test/java/cucumbertest/StepsDefinitionRotation.java
@@ -1,10 +1,10 @@
 package cucumbertest;
-import cucumbertest.Robot;
 import decks.Deck;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import model.board.Board;
+import model.tile.Robot;
 import subcards.Card;
 import subcards.MoveOne;
 import subcards.MoveThree;
