+++ cucumber-test/src/test/java/cucumbertest/StepsDefinitionSelect.java
@@ -1,12 +1,12 @@
 package cucumbertest;
 import static org.junit.Assert.assertEquals;
 
-import cucumbertest.Robot;
 import decks.Deck;
 import decks.SubDeck;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
+import model.tile.Robot;
 import subcards.Card;
 
 public class StepsDefinitionSelect {
